+++ src/modules/auction/service-layer/auction.service.ts
@@ -16,6 +16,7 @@ import { RewardTierNotFoundException } from './exceptions/RewardTierNotFoundExce
 import { RewardTierBadOwnerException } from './exceptions/RewardTierBadOwnerException';
 import { UsersService } from '../../users/users.service';
 import { classToPlain } from 'class-transformer';
+import { UploadResult } from 'src/modules/file-storage/model/UploadResult';
 
 @Injectable()
 export class AuctionService {
@@ -177,12 +178,13 @@ export class AuctionService {
 
   private async validateTierPermissions(userId: number, tierId: number) {
     const tier = await this.rewardTierRepository.findOne({ where: { id: tierId } });
+
     if (!tier) {
-      //return tier not found
+      throw new RewardTierNotFoundException();
     }
 
-    if (userId !== tier.userId) {
-      //return error if missmatch
+    if (tier.userId !== userId) {
+      throw new RewardTierBadOwnerException();
     }
 
     return tier;
@@ -194,21 +196,26 @@ export class AuctionService {
     params: { customDescription: string; tierColor: string },
   ) {
     // const tier = await this.validateTierPermissions(userId, tierId);
-
     // tier.customDescription = params.customDescription ? params.customDescription : tier.customDescription;
     // tier.tierColor = params.tierColor ? params.tierColor : tier.tierColor;
     // await this.rewardTierRepository.save(tier);
     // return tier;
   }
 
-  async updateRewardTierImage(userId: number, tierId: number, file: Express.Multer.File) {
-    // const tier = await this.validateTierPermissions(userId, tierId);
+  async updateRewardTierImage(userId: number, tierId: number, image: Express.Multer.File) {
+    const user = await this.usersService.getById(userId, true);
+    const rewardTier = await this.validateTierPermissions(user.id, tierId);
 
-    // await this.s3Service.uploadDocument(`${file.path}`, `${file.filename}`);
+    let s3Result: UploadResult;
 
-    // tier.tierImageUrl = file.filename;
-    // await this.rewardTierRepository.save(tier);
-    // return tier;
+    if (image) {
+      s3Result = await this.s3Service.uploadDocument(image.path, image.filename);
+      rewardTier.imageUrl = s3Result.url;
+      await this.rewardTierRepository.save(rewardTier);
+      await this.fileSystemService.removeFile(image.path);
+    }
+
+    return classToPlain(rewardTier);
   }
 
   @Transaction()
@@ -281,6 +288,28 @@ export class AuctionService {
     return auction;
   }
 
+  async cancelFutureAuction(userId: number, auctionId: number) {
+    const auction = await this.validateAuctionPermissions(userId, auctionId);
+    let canceled = false;
+    const now = new Date();
+    // TODO: Add more validations if needed
+    if (now < auction.startDate) {
+      await getManager().transaction(async (transactionalEntityManager) => {
+        await transactionalEntityManager.delete(Auction, { id: auctionId });
+        const rewardTiers = await this.rewardTierRepository.find({ auctionId: auction.id });
+        const rewardTiersIdsToDelete = rewardTiers.map((tier) => tier.id);
+        await transactionalEntityManager.delete(RewardTier, { id: In(rewardTiersIdsToDelete) });
+        await transactionalEntityManager.delete(RewardTierNft, { rewardTierId: In(rewardTiersIdsToDelete) });
+      });
+      canceled = true;
+    }
+
+    return {
+      id: auction.id,
+      canceled,
+    };
+  }
+
   async uploadAuctionLandingImages(
     userId: number,
     auctionId: number,
@@ -316,7 +345,6 @@ export class AuctionService {
     const [auctions, count] = await this.auctionRepository.findAndCount({
       where: {
         userId,
-        startDate: MoreThan(now),
       },
       skip: offset,
       take: limit,
@@ -492,7 +520,7 @@ export class AuctionService {
     // }
   }
 
-  async listAuctionsByUser(userId: number, page: number = 1, limit: number = 10) {
+  async listAuctionsByUser(userId: number, page = 1, limit = 10) {
     // const auctionsQuery = this.auctionRepository.createQueryBuilder().where('userId = :userId', { userId });
     // const countQuery = auctionsQuery.clone();
     //
@@ -507,7 +535,7 @@ export class AuctionService {
     // };
   }
 
-  async listAuctionsByUserAndStatus(userId: number, status: string, page: number = 1, limit: number = 10) {
+  async listAuctionsByUserAndStatus(userId: number, status: string, page = 1, limit = 10) {
     const auctionsQuery = this.auctionRepository.createQueryBuilder();
 
     if (status == AuctionStatus.draft) {
@@ -535,7 +563,7 @@ export class AuctionService {
     };
   }
 
-  async listAuctions(page: number = 1, limit: number = 10) {
+  async listAuctions(page = 1, limit = 10) {
     const auctionsQuery = this.auctionRepository.createQueryBuilder();
     const countQuery = auctionsQuery.clone();
 
@@ -550,7 +578,7 @@ export class AuctionService {
     };
   }
 
-  async listAuctionsByStatus(status: string, page: number = 1, limit: number = 10) {
+  async listAuctionsByStatus(status: string, page = 1, limit = 10) {
     const auctionsQuery = this.auctionRepository.createQueryBuilder();
 
     if (status == AuctionStatus.draft) {